0.2.5 - Bugfix: settlement sliders set to max can produce constant drag-sliding noises

+++ 1.6/Source/Core/Comps/WorldComponent_SupplyChain.cs
@@ -375,7 +375,7 @@ namespace FactionColonies.SupplyChain
 
         private static ulong FlowKey(int settlementTile, ushort resourceIndex)
         {
-            return ((ulong)(uint)settlementTile << 16) | resourceIndex;
+            return ((ulong)settlementTile << 16) | resourceIndex;
         }
 
         internal FlowBreakdown GetCachedFlow(WorldSettlementFC settlement, WorldObjectComp_SupplyChain comp, ResourceTypeDef def)
